refactor: authをfeaturesに
