+++ apps/web/src/app/protected/reset-password/page.tsx
@@ -1,9 +1,9 @@
-import { FormMessage, Message } from '@/components/form-message';
-import { SubmitButton } from '@/components/submit-button';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 
 import { resetPasswordAction } from '@/app/actions';
+import { FormMessage, Message } from '@/features/auth/form-message';
+import { SubmitButton } from '@/features/auth/submit-button';
 
 export default async function ResetPassword(props: { searchParams: Promise<Message> }) {
   const searchParams = await props.searchParams;
